requirements.txt
# ==================== 核心依赖 ====================
# PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Web 框架
fastapi>=0.100.0
uvicorn>=0.22.0
aiohttp>=3.8.0
aiohttp-cors>=0.7.0
flask>=3.0.0
flask-sockets>=0.2.1

# WebRTC
aiortc>=1.5.0

# ==================== AI/ML ====================
# Transformers
transformers>=4.30.0
accelerate>=0.20.0
diffusers>=0.21.0
omegaconf>=2.3.0

# OpenAI
openai>=1.0.0

# DashScope (阿里云)
dashscope>=1.25.0

# ==================== 音频处理 ====================
librosa>=0.10.0
soundfile>=0.12.0
resampy>=0.4.0
python_speech_features>=0.6
edge-tts>=6.1.0
websocket-client>=1.6.0
websockets>=12.0

# ==================== 图像/视频处理 ====================
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
ffmpeg-python>=0.2.0

# ==================== 人脸处理 ====================
face-alignment>=1.3.5
dlib>=19.24.0

# ==================== 数据处理 ====================
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# ==================== 可视化 ====================
matplotlib>=3.7.0
tensorboardX>=2.6.0
rich>=13.0.0

# ==================== 工具库 ====================
tqdm>=4.65.0
requests>=2.31.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
configargparse>=1.5.0
packaging>=23.0

# ==================== 数据库/存储 ====================
chromadb>=0.4.0
sentence-transformers>=2.2.0

# ==================== 其他 ====================
einops>=0.6.0
numba>=0.57.0
lpips>=0.1.3
trimesh>=3.21.0
PyMCubes>=0.1.4
ninja>=1.11.0
typeguard==2.13.3
onnxruntime-gpu>=1.15.0
certifi>=2023.0.0

# ==================== 开发依赖 ====================
pytest>=7.0.0
black>=23.0.0
ruff>=0.0.280
